third_party/blink/web_tests/flag-specific/highdpi/fast/forms/color/color-picker-appearance-color-well-down-accelerated-keyboard-navigation-from-top-left-corner-expected.txt

This is a testharness.js-based test.
FAIL Color picker: Moving the color well's selection ring using accelerated keyboard movement. assert_equals: expected "69" but got "66"
Harness: the test ran to completion.
